+++ src/impls/buf_bit_writer.rs
@@ -25,7 +25,7 @@ use mem_dbg::{MemDbg, MemSize};
 /// instantaneous codes, but the casual user should be happy with the default
 /// value. See [`WriteParams`] for more details.
 ///
-/// The supported range for [writing unary codes](Self:write_unary) is
+/// The supported range for [writing unary codes](BitWrite::write_unary) is
 /// [0 . . 2⁶⁴ – 1).
 ///
 /// For additional flexibility, this structures implements [`std::io::Write`].